Restructured files and made a simple footer
