+++ src/components/ImageCarousel/index.tsx
