+++ src/pages/Homepage/Homepage.module.css
@@ -3,7 +3,6 @@
   flex-direction: column;
   align-items: center;
   margin-inline: 15%;
-  padding-bottom: 100px;
 }
 
 .pageCounter {
